# Optimize access_log_record.go