Coalesce log line printing in a single Sprintf and memoize the result so that we don't build each log line twice
Three small functional changes:
- Missing values are replaced by `-` instead of the verbose `Missing<name-of-field>` placeholder
- The request duration does not print pointless 0s after the decimal separator
- Extra headers values are properly escaped using strconv.Quote

With the changes to the benchmark included in the PR (github.com/cloudfoundry/gorouter/access_log/file_and_loggregator_access_logger_test.go) we get, without the patch:
```
  Ran 500 samples:
  writeTime:
    Fastest Time: 0.009s
    Slowest Time: 0.046s
    Average Time: 0.011s ± 0.005s
```

and, with the patch:
```
  Ran 500 samples:
  writeTime:
    Fastest Time: 0.004s
    Slowest Time: 0.036s
    Average Time: 0.005s ± 0.003s
```