@shalako 2015-09-16
Hello @CAFxX, thank you for your contribution. Was this discussed on cf-dev first? In the future, please ping me with a proposal before doing the work. Thanks.

- Please provide example log output for before and after
- I'm skeptical that removing `Missing<name of field>` is a positive change. This format is self-documenting and more immediately meaningful. What is driving this change? Some performance improvement by the removal of a few characters? If we are going to go forward with this change, we'll need discoverable documentation for the log format, that describes the order and meaning of fields in the log line. This should be added here: https://github.com/cloudfoundry/gorouter#logs