@CAFxX 2015-09-16
@shalako it wasn't discussed on cf-dev, but briefly discussed with @fraenkel in #92.

The primary driver was simply optimizing the code without changing any existing behavior. But during the refactoring I noticed some issues:

- Most fields, when missing, are replaced by a `-`. Just three of them are replaced by `Missing<NameOfField>`. For consistency I replaced them with `-`. This change is opinionated and can be reverted, if you think it is appropriate.
- The request duration field was formatted with 9 digits after the decimal separator in all cases. So 60 would be formatted as `60.000000000`, 60.1 would be `60.100000000` and 60.000000001 would be `60.000000001`. I changed it to print decimal digits only if they are meaningful (i.e. `60`, `60.1` and `60.000000001` for the cases above). Again, opinionated change that can be reverted if needed.

How about I rescope this PR for the optimization part (i.e. no format changes), and leave the opinionated changes for another day?