@shalako 2015-09-16
@CAFxX Thank you for the clarification. 

No need to rescope yet. I'll pull the story into the backlog to have the team review.