@msmykowski 2015-10-05
@CAFxX We were curious why you decided to write the access log record twice in the benchmark test?

https://github.com/rakutentech/gorouter/blob/4d67340cfc283ceec37cae61c8099aff42b67533/access_log/file_and_loggregator_access_logger_test.go#L85-L88