@crhino 2015-11-06
This PR was pulled in on c068f24e4089f6807c389a072cf0d8767db6ac66, but wasn't merged so no automatic closing of this PR happened.

Closing now.